requirements.txt
flask
openai
pdfplumber
python-docx
